+++ CHANGES.rst
@@ -31,3 +31,5 @@ Changelog
 
 - Added the ability to specify inputs by glob syntax, and also specify negative
   images with no annotations.
+- Added utility functions for images (image_to_numpyarray, remove_alpha_channel)
+  and for masks (mask_to_polygon, polygon_to_minrect) in module wai.annotations.image_utils